chore(scripts): remove storybook/runner as its not needed anymore
